+++ src/common/logging.h
@@ -170,5 +170,5 @@ void checkedLog(std::string logger, std::string level, Args... args) {
   }
 }
 
-void createLoggers(const marian::Config* options = nullptr);
+std::vector<Logger> createLoggers(const marian::Config* options = nullptr);
 void switchtoMultinodeLogging(std::string nodeIdStr);