+++ src/common/types.h
@@ -17,7 +17,17 @@
 #include <type_traits>
 
 #ifndef __CUDACC__ // NVCC is very unreliable when it comes to CPU intrinsics, we hide them completely from NVCC-compiled code
+
+#ifdef __AVX__
 #include <immintrin.h>
+#elif __SSE__
+#include <xmmintrin.h>
+#endif
+
+#if defined(__ARM_NEON) || defined(__ARM_NEON__)
+#include <arm_neon.h>
+#endif
+
 #endif
 
 #ifdef __CUDACC__ // nvcc is compiling this code
@@ -163,6 +173,43 @@ struct intgemm8 {
 
 #ifndef __CUDACC__ // vectorized types not available from .cu files
 
+
+#if defined(__ARM_NEON) || defined(__ARM_NEON__)
+// The following struct fills this structure for ARM with NEON SIMD, changing
+// __m128 and _mm_set1_ps with the equivalents on ARM-NEON.
+struct float32x4 {
+private:
+    // NEON uses 128-bit SIMD registers, same as SSE. We are copying this class
+    // and locally aliasing __m128 to float32x4_t, which is the NEON
+    // equivalent.
+   using __m128 = float32x4_t;
+  __m128 f_;
+
+public:
+  float32x4() {}
+  float32x4(const __m128& f) : f_(f) {}
+  // __m128 _mm_set1_ps(float) copies value into all slots, vdupq_n_f32 is it's
+  // NEON equivalent.
+  float32x4(const float& f) : f_(vdupq_n_f32(f)) {} 
+
+  operator const __m128&() const { return f_; }
+  operator __m128&() { return f_; }
+
+  float operator[] (size_t i) const {
+    return *(((float*)&f_) + i); // potentially undefined, but efficient. In practice __m128 is an array of floats
+  }
+
+  friend std::ostream& operator<<(std::ostream& out, float32x4 f4) {
+    float* a = (float*)&f4;
+    out << "[" << a[0];
+    for(int i = 1; i < 4; i++)
+      out << " " << a[i];
+    out << "]";
+    return out;
+  }
+};
+
+#else 
 // @TODO: check what intrinsics are actually available.
 struct float32x4 {
 private:
@@ -189,6 +236,8 @@ public:
     return out;
   }
 };
+
+#endif
 
 // @TODO: consider how code can be shared via templating
 #ifdef __AVX__