+++ src/functional/operators.h
@@ -213,7 +213,16 @@ struct Ops<double> {
 // __CUDA_ARCH__ is defined when compiling device (GPU) code
 #ifndef __CUDACC__
 
+#if defined(__ARM_NEON) || defined(__ARM_NEON__)
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#include "3rd_party/simd_utils/simd_utils.h"
+#pragma GCC diagnostic pop
+#else
 #include "3rd_party/sse_mathfun.h"
+#include <immintrin.h>
+#endif
+
 
 namespace marian {
 namespace functional {
@@ -269,6 +278,10 @@ struct Ops<float32x4> {
   static inline float32x4 round(const float32x4& x)  { return _mm_round_ps(x, _MM_FROUND_TO_NEAREST_INT); }
   static inline float32x4 floor(const float32x4& x)  { return _mm_floor_ps(x); }
   static inline float32x4 ceil(const float32x4& x)   { return _mm_ceil_ps(x); }
+#else
+  static inline float32x4 round(const float32x4& x)  { return loop4(Ops<float>::round, x); }
+  static inline float32x4 floor(const float32x4& x)  { return loop4(Ops<float>::floor, x); }
+  static inline float32x4 ceil(const float32x4& x)  { return loop4(Ops<float>::ceil, x); }
 #endif
   static inline float32x4 add(const float32x4& x, const float32x4& y) { return _mm_add_ps(x, y); }
   static inline float32x4 sub(const float32x4& x, const float32x4& y) { return _mm_sub_ps(x, y); }