+++ src/tensors/cpu/backend.h
@@ -33,6 +33,9 @@ protected:
       setInt16(true);
     } else if (gemmPrecision == "int8") {
       setInt8(true);
+    } else if (gemmPrecision == "int8Alpha") {
+      setInt8(true);
+      setPrecomputedAlpha(true);
     } else if (gemmPrecision == "int8shift") {
       setInt8(true);
       setShifted(true);
@@ -75,14 +78,27 @@ public:
   void setInt8(bool optimize) override { int8_ = optimize; }
   bool isInt8() override { return int8_; }
 
-  void setShifted(bool shifted) override { shifted_ = shifted; }
+  void setShifted(bool shifted) override { 
+#if (defined(__arm__) || defined(__aarch64__))
+      LOG(info, "gemm-precision: *shifted* is not available on ARM; Setting to false.");
+      shifted_ = false;
+#else
+      shifted_ = shifted; 
+#endif
+  }
   bool isShifted() override { return shifted_; }
 
   void setShiftedAll(bool shiftedAll) override {
+#if (defined(__arm__) || defined(__aarch64__))
+      LOG(info, "gemm-precision: *shifted* is not available on ARM; Setting to false.");
+      shiftedAll_ = false;
+      shifted_ = false;
+#else
     shiftedAll_ = shiftedAll;
     if (shiftedAll_) {
       shifted_ = true;
     }
+#endif
   }
 
   bool isShiftedAll() override {